% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addressToCoord.R
\name{addressToCoord}
\alias{addressToCoord}
\title{Address to Coordinate}
\usage{
addressToCoord(address, source = "google")
}
\arguments{
\item{address}{A text string containing the address.}

\item{source}{See ?ggmap::geocode}
}
\value{
A vector of longitude and latitude values.
}
\description{
This function maps an address (text) to the latitude and longitude
coordinates.
}

