% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assignDirectory.R
\name{assignDirectory}
\alias{assignDirectory}
\title{Get Directory}
\usage{
assignDirectory()
}
\value{
No value is returned, but workingDir and savingDir are written to
  the global environment.
}
\description{
This function uses the information at Sys.info() to assign a working
directory and saving directory.  The names of the two objects created are
workingDir and savingDir, and they are assigned to the global environment so
as to not require returning/assigning by this function.
}

