% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cleanImb.R
\name{cleanImb}
\alias{cleanImb}
\title{Clean Immobiliare data}
\usage{
cleanImb(data)
}
\arguments{
\item{data}{The Immobiliare data.table}
}
\value{
The same object as passed, but after some data cleaning.
}
\description{
This function cleans up the Immobiliare data by standardizing columns,
converting variables to appropriate formats, etc.
}

