% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cleanMioAffitto.R
\name{cleanMioAffitto}
\alias{cleanMioAffitto}
\title{Clean Mio Affitto data}
\usage{
cleanMioAffitto(data)
}
\arguments{
\item{data}{The Mio Affitto data.table}
}
\value{
The same object as passed, but after some data cleaning.
}
\description{
This function cleans up the Mio Affitto data by standardizing columns,
converting variables to appropriate formats, etc.
}

