% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compileScrapedHistory.R
\name{compileScrapedHistory}
\alias{compileScrapedHistory}
\title{Compile scraped history}
\usage{
compileScrapedHistory(dir = paste0(savingDir,
  "historical/data_by_cap/scraped_data/"))
}
\arguments{
\item{dir}{The directory where all the individual files are stored.}
}
\value{
No value is returned, but a file called allHistoricalData.csv is
  created in dir.
}
\description{
This function takes the scraped history files for average house prices by
CAP from immobiliare and creates one main data file for plotting.
}

