% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getAgenzieDetailsImmobiliare.R
\name{getAgenzieDetailsImmobiliare}
\alias{getAgenzieDetailsImmobiliare}
\title{Get contact details of housing agenices listed on immobiliare.it}
\usage{
getAgenzieDetailsImmobiliare(numPages)
}
\arguments{
\item{numPages}{The number of pages listing agencies from immobiliare.it}
}
\value{
A data.table listing the agencies and contact details
}
\description{
This function takes the number of pages from the first page of the list of agencies from
immobiliare.it, and a data frame with the list of agencies and contact details.
}

