% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getAgenzieNumImmobiliare.R
\name{getAgenzieNumImmobiliare}
\alias{getAgenzieNumImmobiliare}
\title{Get number of pages listing housing agencies in Rome from immobiliare.it}
\usage{
getAgenzieNumImmobiliare(url)
}
\arguments{
\item{url}{The url of the first page of agency listings}
}
\value{
A one element list with number of pages
}
\description{
This function takes the url of the first page of the list of agencies from
immobiliare.it, and returns the number of pages listing agencies
}

