% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getCasaMainPages.R
\name{getCasaMainPages}
\alias{getCasaMainPages}
\title{Create urls to feed to get PropertyUrlsCasa}
\usage{
getCasaMainPages(type = "vendita")
}
\arguments{
\item{type}{A character string, either "vendita" or "affitto", indicating
  which type of urls should be scraped.}
}
\value{
A character vector containing the urls for all the individual
  MAIN pages for listing
}
\description{
This function gets the urls for each of the detailed, property listing
pages.
}

