% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getMetroData.R
\name{getMetroData}
\alias{getMetroData}
\title{Get Metro Data}
\usage{
getMetroData(file = "~/../Dropbox/romeHouseData/")
}
\arguments{
\item{file}{The file location where the metro data should be saved.}
}
\value{
No file is returned, but the metro station locations are saved to the provided directory.
}
\description{
This function pulls the locations of the metro stops using the names of the metro stations as provided by the ATAC website.
}

