% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getNumPagesCasa.R
\name{getNumPagesCasa}
\alias{getNumPagesCasa}
\title{Get Number of Pages for Casa.it}
\usage{
getNumPagesCasa(url)
}
\value{
An integer giving the number of pages.
}
\description{
This function returns the number of listing pages available on the
casa.it housing sale site.
}

