% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getNumPagesImmobiliare.R
\name{getNumPagesImmobiliare}
\alias{getNumPagesImmobiliare}
\title{Get Number of Pages for Immobiliare}
\usage{
getNumPagesImmobiliare()
}
\value{
An integer giving the number of pages.
}
\description{
This function returns the number of listing pages available on the
immobiliare house sale site.
}

