% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getNumPagesMioAffitto.R
\name{getNumPagesMioAffitto}
\alias{getNumPagesMioAffitto}
\title{Get Number of Pages (Mio Affitto)}
\usage{
getNumPagesMioAffitto()
}
\value{
An integer giving the number of pages.
}
\description{
This function returns the number of listing pages available on the
immobiliare house sale site.
}

