% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getPropertyDetailsCasa.R
\name{getPropertyDetailsCasa}
\alias{getPropertyDetailsCasa}
\title{Get Property Details for Casa.it #copied from immobiliare function}
\usage{
getPropertyDetailsCasa(url)
}
\arguments{
\item{url}{The url of the property listing.}
}
\value{
A data.table (with one row) containing the variables available for
  this dataset.
}
\description{
This function takes the url for a listing page and extracts the relevant
data from it (i.e. square meters, bathrooms, rooms, price, elevator,
restructured, ...)
}

