% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getPropertyDetailsImmobiliare.R
\name{getPropertyDetailsImmobiliare}
\alias{getPropertyDetailsImmobiliare}
\title{Get Property Details for Immobiliare}
\usage{
getPropertyDetailsImmobiliare(url)
}
\arguments{
\item{url}{The url of the property listing.}
}
\value{
A data.table (with one row) containing the variables available for
  this dataset.
}
\description{
This function takes the url for a listing page and extracts the relevant
data from it (i.e. square meters, bathrooms, rooms, price, elevator,
restructured, ...)
}

