% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getPropertyDetailsMioAffitto.R
\name{getPropertyDetailsMioAffitto}
\alias{getPropertyDetailsMioAffitto}
\title{Get Property Details for Mio Affitto}
\usage{
getPropertyDetailsMioAffitto(url)
}
\arguments{
\item{url}{The url of the property listing.}
}
\value{
A data.table (with one row) containing the variables available for
  this dataset.
}
\description{
This function takes the url for a listing page and extracts the relevant
data from it (i.e. square meters, bathrooms, rooms, price, elevator,
restructured, ...)
}
\examples{
\dontrun{
getPropertyDetailsMioAffitto("http://www.mioaffitto.it/affitto_appartamento_roma/tiburtina-appartamento-in-affitto_2892995.html")
}
}

