% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getPropertyUrlsCasa.R
\name{getPropertyUrlsCasa}
\alias{getPropertyUrlsCasa}
\title{Get Property urls for Casa.IT in Rome}
\usage{
getPropertyUrlsCasa(numPages, type = "vendita")
}
\arguments{
\item{numPages}{The number of pages which should be loaded (and 15 urls are
generally scraped per page).}

\item{type}{A character string, either "vendita" or "affitto", indicating
  which type of urls should be scraped.}
}
\value{
A character vector containing the urls for all the individual
  listings.
}
\description{
This function gets the urls for each of the detailed, property listing
pages.
}

