% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getPropertyUrlsImmobiliare.R
\name{getPropertyUrlsImmobiliare}
\alias{getPropertyUrlsImmobiliare}
\title{Get Property Urls for Immobiliare}
\usage{
getPropertyUrlsImmobiliare(numPages, type = "vendita")
}
\arguments{
\item{numPages}{The number of pages which should be loaded (and 15 urls are
generally scraped per page).  If numPages is more than the number of pages
available (likely something in the ~2000 range), this function ends at the
last one (with a warning).}

\item{type}{One of "vendita" or "affitto".  Specifies which types of urls
  should be scraped.}
}
\value{
A character vector containing the urls for all the individual
  listings.
}
\description{
This function gets the urls for each of the detailed, property listing
pages.
}

