% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getPropertyUrlsMioAffitto.R
\name{getPropertyUrlsMioAffitto}
\alias{getPropertyUrlsMioAffitto}
\title{Get Property urls for Mio Affitto}
\usage{
getPropertyUrlsMioAffitto(numPages)
}
\arguments{
\item{numPages}{The number of pages which should be loaded (and 15 urls are
  generally scraped per page).  If numPages is more than the number of pages
  available (likely something in the ~2000 range), this function ends at the
  last one (with a warning).}
}
\value{
A character vector containing the urls for all the individual
  listings.
}
\description{
This function gets the urls for each of the detailed, property listing
pages.
}

