% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotData.R
\name{plotData}
\alias{plotData}
\title{Plot Data}
\usage{
plotData(data)
}
\arguments{
\item{data}{The scraped dataset.}
}
\value{
No object is returned, but a plot is generated.
}
\description{
This function takes one of the data.frames from the scraped sites and plots
it.
}

