% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pullGridAddresses.R
\name{pullGridAddresses}
\alias{pullGridAddresses}
\title{Pull Grid Addresses}
\usage{
pullGridAddresses(width = runif(1, 0, 0.1), resolution = 50,
  shift = runif(2, -0.02, 0.02))
}
\arguments{
\item{width}{The width of the window around Rome}

\item{resolution}{A numeric value specifying how large the grid should be
(resolution x resolution points).}

\item{shift}{A numeric vector of length two that specifies how the window
  should be shifted.}
}
\value{
No object is returned, but the address file is updated.
}
\description{
This function steps up a grid of points around Rome and pulls addresses for
all the points in the grid.  These addresses are then added to the address
file.
}

