% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pullNewAddresses.R
\name{pullNewAddresses}
\alias{pullNewAddresses}
\title{Pull New Addresses}
\usage{
pullNewAddresses()
}
\value{
No object is returned, but saved files are updated with addresses
  (if missing).
}
\description{
This function searches through the current data files and adds on missing
addresses (when the longitude and latitude are available).
}

