% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/substrRight.R
\name{substrRight}
\alias{substrRight}
\title{Extract n characters from the right of a string}
\usage{
substrRight(x, n)
}
\arguments{
\item{x}{character vector}

\item{n}{is the number of characters to extract}
}
\value{
the n characters from the right of the string
}
\description{
This function takes a string, and a number and extracts the number characters
from the right of the string
Taken from http://stackoverflow.com/questions/7963898/extracting-the-last-n-characters-from-a-string-in-r
}

