% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getPropertyUrlsCasa.R
\name{getPropertyUrlsCasa}
\alias{getPropertyUrlsCasa}
\title{Get Property urls for Casa.IT in Rome}
\usage{
getPropertyUrlsCasa(url)
}
\arguments{
\item{url}{is the starting url of the properties to be scraped}
}
\value{
A character vector containing the urls for all the individual
  listings.
}
\description{
This function gets the urls for each of the detailed, property listing
pages.
}

