% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getGridAddresses.R
\name{pullGridAddresses}
\alias{pullGridAddresses}
\title{Pull Grid Addresses}
\usage{
pullGridAddresses(width = 0.1, shift = 0, resolution = 3)
}
\arguments{
\item{width}{width}

\item{shift}{shift}

\item{resolution}{resolution}
}
\value{
No object is returned, but the address file is updated.
}
\description{
Pull Grid Addresses
}

